move bq_test to db.py
